producers/requirements.txt
Faker
kafka-python
ipykernel
requests